Update CompareTool and fix float precision issues
Move all GS and ImageMagick handling to the separate class

DEVSIX-4067

+++ kernel/src/main/java/com/itextpdf/kernel/pdf/canvas/parser/data/TextRenderInfo.java
@@ -479,7 +479,8 @@ public class TextRenderInfo extends AbstractRenderInfo {
         checkGraphicsState();
         if (singleCharString) {
             float[] widthAndWordSpacing = getWidthAndWordSpacing(string);
-            return (widthAndWordSpacing[0] * gs.getFontSize() + gs.getCharSpacing() + widthAndWordSpacing[1]) * gs.getHorizontalScaling() / 100f;
+            return (float)(((double)widthAndWordSpacing[0] * (double)gs.getFontSize() + (double)gs.getCharSpacing()
+                    + (double)widthAndWordSpacing[1]) * (double)gs.getHorizontalScaling() / 100f);
         } else {
             float totalWidth = 0;
             for (PdfString str : splitString(string)) {
